All Controllers checked, authorisation included

+++ MeFit.API/Controllers/ProgramsController.cs
@@ -10,10 +10,11 @@ using MeFit.DAL.Models.Domain;
 using MeFit.DAL.Models.DTOs.Program;
 using AutoMapper;
 using System.Net.Mime;
+using Microsoft.AspNetCore.Authorization;
 
 namespace MeFit.API.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("api")]
     [ApiController]
     public class ProgramsController : ControllerBase
     {
@@ -26,14 +27,16 @@ namespace MeFit.API.Controllers
             _mapper = mapper;
         }
 
-        // GET: api/Programs
+        // GET: api/programs
         /// <summary>
         /// Gets all programs with workouts IDs
         /// </summary>
         /// <returns>List of all programs with workouts IDs</returns>
         /// <response code="200">Returns all programs</response>
         /// <response code="204">No programs found</response>
-        [HttpGet]
+        /// <response code="401">Not authorized</response>
+        [HttpGet("programs")]
+        [Authorize]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<IEnumerable<ProgramReadDTO>>> GetPrograms()
@@ -48,20 +51,22 @@ namespace MeFit.API.Controllers
             return Ok(programs);
         }
 
-        // GET: api/Programs/5
+        // GET: api/program/5
         /// <summary>
         /// Gets program by ID
         /// </summary>
         /// <param name="id">ID of a program</param>
         /// <returns>Program with workouts IDs</returns>
         /// <response code="200">Returns a program</response>
+        /// <response code="401">Not authorized</response>
         /// <response code="404">No program found</response>
-        [HttpGet("{id}")]
+        [HttpGet("program/{id}")]
+        [Authorize]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<ProgramReadDTO>> GetProgramById([FromRoute] int id)
         {
-            var program = await _context.Programs.Include(p=> p.Workouts).FirstOrDefaultAsync(p => p.Id == id);
+            var program = await _context.Programs.Include(p => p.Workouts).FirstOrDefaultAsync(p => p.Id == id);
 
             if (program == null)
             {
@@ -71,7 +76,7 @@ namespace MeFit.API.Controllers
             return Ok(_mapper.Map<ProgramReadDTO>(program));
         }
 
-        // PUT: api/Programs/5/AddWorkouts -------------CONTRIBUTOR ONLY!!!!! -------------
+        // PATCH: api/program/5/AddWorkouts -------------CONTRIBUTOR ONLY!!!!! -------------
         /// <summary>
         /// Add workouts to program
         /// </summary>
@@ -79,9 +84,11 @@ namespace MeFit.API.Controllers
         /// <param name="workoutIds">List of workouts IDs to be added</param>
         /// <response code="204">Successfully added workouts to program</response>
         /// <response code="404">No program found</response>
-        [HttpPatch("{id}/AddWorkouts")]
+        /// <response code="500">Internal Server Error</response>
+        [HttpPatch("program/{id}/AddWorkouts")]
+        [Authorize(Roles = "contributor, administrator")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
-        public async Task<IActionResult> PutProgram([FromRoute] int id, [FromBody] List<int> workoutIds)
+        public async Task<IActionResult> AddWorkoutsToProgram([FromRoute] int id, [FromBody] List<int> workoutIds)
         {
             var program = await _context.Programs.Include(p => p.Workouts).FirstOrDefaultAsync(p => p.Id == id);
 
@@ -105,49 +112,62 @@ namespace MeFit.API.Controllers
             {
                 await _context.SaveChangesAsync();
             }
-            catch (DbUpdateConcurrencyException)
+            catch
             {
-                if (!ProgramExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
             return NoContent();
         }
 
-        // POST: api/Programs -------------CONTRIBUTOR ONLY!!!!! -------------
+        // POST: api/program -------------CONTRIBUTOR ONLY!!!!! -------------
         /// <summary>
         /// Creates new program
         /// </summary>
         /// <param name="newProgram">New program info</param>
         /// <returns>A newly created program</returns>
         /// <response code="201">Successfully created program</response>
-        [HttpPost]
+        /// <response code="401">Not authorized</response>
+        /// <response code="403">Not allowed(not having the necessary permissions)</response>
+        /// <response code="500">Internal Server Error</response>
+        [HttpPost("program")]
+        [Authorize(Roles = "contributor, administrator")]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status201Created)]
         public async Task<ActionResult<ProgramReadDTO>> PostProgram([FromBody] ProgramCreateDTO newProgram)
         {
-            var domainProgram = _mapper.Map<Program>(newProgram);
-            _context.Programs.Add(domainProgram);
-            await _context.SaveChangesAsync();
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
 
-            return CreatedAtAction("GetProgramById", new { id = domainProgram.Id }, _mapper.Map <ProgramReadDTO>(newProgram));
+            var domainProgram = _mapper.Map<Program>(newProgram);
+            domainProgram.OwnerId = userId;
+            _context.Programs.Add(domainProgram);
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
+
+            return CreatedAtAction("GetProgramById", new { id = domainProgram.Id }, _mapper.Map<ProgramReadDTO>(newProgram));
         }
 
-        // DELETE: api/Programs/5 ------------- CONTRIBUTOR ONLY!!!!! -------------
+        // DELETE: api/program/5 ------------- CONTRIBUTOR ONLY!!!!! -------------
         /// <summary>
         /// Deletes a program
         /// </summary>
         /// <param name="id">ID of a program</param>
         /// <response code="204">Successfully deleted program</response>
+        /// <response code="401">Not authorized</response>
+        /// <response code="403">Not allowed(not having the necessary permissions)</response>
         /// <response code="404">No program found</response>
-        [HttpDelete("{id}")]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        /// <response code="500">Internal Server Error</response>
+        [HttpDelete("program/{id}")]
+        [Authorize(Roles = "contributor, administrator")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<IActionResult> DeleteProgram([FromRoute] int id)
         {
             var program = await _context.Programs.FindAsync(id);
@@ -156,15 +176,35 @@ namespace MeFit.API.Controllers
                 return NotFound();
             }
 
-            _context.Programs.Remove(program);
-            await _context.SaveChangesAsync();
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
+            if (program.OwnerId != userId)
+            {
+                return Forbid();
+            }
+
+            try
+            {
+                _context.Programs.Remove(program);
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return NoContent();
         }
 
-        private bool ProgramExists(int id)
+        private string TakeUserNameFromToken()
         {
-            return _context.Programs.Any(e => e.Id == id);
+            var username = User.Claims.FirstOrDefault(c => c.Type == "preferred_username");
+            return username.Value;
+        }
+        private async Task<int> TakeIdFromUser(string usernameFromToken)
+        {
+            var id = (await _context.Users.FirstOrDefaultAsync(x => x.Username == usernameFromToken)).Id;
+            return id;
         }
     }
 }
